libnm-core: return variants with full reference when parsing attributes
When variants get added to the container (hash table) their floating
reference must be sinked, so that the container assumes the ownership.

Previously we returned floating GVariants in the hash table and the
caller had to call g_variant_ref_sink() on the values to turn the
floating reference into a full one. The documentation didn't specify
this and I think this wasn't intended.

This is a change in behavior of public API, and therefore I'm not sure
we can do this change now, but I believe the previous code was buggy.

Fixes: 93b3a478bb5236f37be3d6d3414755ae17cbec04