device: clear dad6_failed_addrs in _cleanup_ip_pre()
We also cancel the idle handler

  nm_clear_g_source (&priv->queued_ip_config_id_x[IS_IPv4])

which means, nobody is going to process these addresses (at least
for the moment).

dad6_failed_addrs is just keeping track to look at addresses
that might be interesting for checking DAD. If we are going
to deconfigure the device, we can forget about these reminders.